ffv-meshviewer-filter: Work around inconsistent graph data
For some reason the graph+node data of FFC has become inconsistent. More
nodes are in the graph data then there are in the nodes data. Thus we
cannot assume that all nodes in the graph data can be checked for there
validity using an information generated using the node data.

Signed-off-by: Sven Eckelmann <sven@narfation.org>